+++ CHANGELOG.rst
@@ -59,6 +59,7 @@ Change Log
 - [IMPROVED] force class attributes to be numpy arrays of proper types when the 
   classes are initialized from the backend.
 - [IMPROVED] some (slight) speed improvments when comparing actions or deep copying objects
+- [IMPROVED] the way the "grid2op compat" mode is handled
 
 [1.10.1] - 2024-03-xx
 ----------------------